# components/onboarding/api/swagger.yaml
basePath: /
definitions:
...
  Error:
    description: Standardized error response format used across all API endpoints
      for error situations. Provides structured information about errors including
      codes, messages, and field-specific validation details.
    properties:
      code:
        description: |-
          Error code identifying the specific error condition
          example: ERR_INVALID_INPUT
          maxLength: 50
        example: ERR_INVALID_INPUT
        maxLength: 50
        type: string
      entityType:
        description: |-
          Optional type of entity associated with the error
          example: Organization
          maxLength: 100
        example: Organization
        maxLength: 100
        type: string
      fields:
        additionalProperties:
          type: string
        description: |-
          Optional detailed field validations for client-side handling
          example: {"name": "Field 'name' is required"}
        example:
          '{"name"': ' "Field ''name'' is required"}'
        type: object
      message:
        description: |-
          Detailed error message explaining the issue
          example: The request contains invalid fields. Please check the field 'name' and try again.
          maxLength: 500
        example: The request contains invalid fields. Please check the field 'name'
          and try again.
        maxLength: 500
        type: string
      title:
        description: |-
          Short, human-readable error title
          example: Bad Request
          maxLength: 100
        example: Bad Request
        maxLength: 100
        type: string
    type: object
...
  Pagination:
    description: Pagination is the struct designed to store the pagination data of
      an entity list.
    properties:
      items: {}
      limit:
        example: 10
        type: integer
      next_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
      page:
        example: 1
        type: integer
      prev_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
    type: object
...
host: localhost:3000
info:
  contact:
    name: Discord community
    url: https://discord.gg/DnhqKwkGv3
  description: This is a swagger documentation for the Midaz Ledger API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Midaz Onboarding API
  version: v1.48.0
paths:
...
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/accounts:
    get:
      description: Returns a paginated list of accounts within the specified ledger,
        optionally filtered by metadata, date range, and other criteria
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: JSON string to filter accounts by metadata fields
        in: query
        name: metadata
        type: string
      - default: 10
        description: Maximum number of records to return per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page number for pagination
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Filter accounts created on or after this date (format: YYYY-MM-DD)'
        in: query
        name: start_date
        type: string
      - description: 'Filter accounts created on or before this date (format: YYYY-MM-DD)'
        in: query
        name: end_date
        type: string
      - description: Sort direction for results based on creation date
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved accounts list
          schema:
            allOf:
            - $ref: '#/definitions/Pagination'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/Account'
                  type: array
                limit:
                  type: integer
                page:
                  type: integer
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Organization or ledger not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: List all accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Creates a new account within the specified ledger. Accounts represent
        individual financial entities like bank accounts, credit cards, or expense
        categories.
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Account details including name, type, asset code, and optional
          parent account, portfolio, segment, and metadata
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/CreateAccountInput'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created account
          schema:
            $ref: '#/definitions/Account'
        "400":
          description: Invalid input, validation errors
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Organization, ledger, parent account, portfolio, or segment
            not found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: 'Conflict: Account with the same alias already exists'
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Create a new account
      tags:
      - Accounts
...
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/accounts/alias/{alias}:
    get:
      description: Returns detailed information about an account identified by its
        alias within the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Account alias (e.g. @person1)
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved account
          schema:
            $ref: '#/definitions/Account'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Account with the specified alias, ledger, or organization not
            found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Retrieve an account by alias
      tags:
      - Accounts
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/accounts/external/{code}:
    get:
      description: Returns detailed information about an account identified by its
        alias within the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Account External Code (e.g. BRL)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved account
          schema:
            $ref: '#/definitions/Account'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Account with the specified alias, ledger, or organization not
            found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Retrieve an account by alias
      tags:
      - Accounts
...
swagger: "2.0"